rename: seperating folders and moving files
src폴더 하위에 있는 폴더들의 역할 정리
1. assets: static resource들을 담아두는 폴더
2. components: 재사용이 가능한 컴포넌트들
3. constants: 상수를 담아두는 폴더
4. contexts: Context API 관련 파일
5. pages: React Routing이 부르는 화면단위의 파일
6. services: API 관련 파일
7. styles: 스타일 관련 파일
