+++ src/App.jsx
@@ -2,9 +2,8 @@ import { react, useState } from 'react'
 import './App.css'
 import 'bootstrap/dist/css/bootstrap.min.css';
 
-import Header from './components/commons/Header';
-
-import Landing from './components/landing/Landing'
+import Header from './pages/Header';
+import Landing from './pages/Landing'
 
 function App() {
   const [count, setCount] = useState(0)
