+++ src/components/landing/Courses.jsx
@@ -1,6 +1,6 @@
 import PropTypes from 'prop-types';
 import CourseCardItem from './CourseCardItem';
-import styles from '@/components/commons/LandingStyles.module.css'
+import styles from '@/styles/LandingStyles.module.css'
 
 const Courses = (props) => {
